wishlists

+++ amazen_server/Repositories/ProductRepository.cs
@@ -53,10 +53,10 @@ namespace amazen_server.Repositories
             _db.Execute(sql, data);
         }
 
-        internal void Delete(int id)
+        internal void Delete(int id, Profile userInfo)
         {
-            string sql = "delete from products where id = @id";
-            _db.Execute(sql, new { id });
+            string sql = "delete from products where id = @id and creatorId = @creatorId";
+            _db.Execute(sql, new { id, creatorId = userInfo.Id });
         }
     }
 }
